# PE/lab04/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "API приложения для хранения файлов",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8099"
        }
    ],
    "paths": {
        "/api/v1/register": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Create User",
                "operationId": "create_user_api_v1_register_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "detail": "User not found"
                    },
                    "400": {
                        "description": "Bad Request",
                        "detail": "Invalid user data"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "detail": "Internal error"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/{user_id}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get User",
                "operationId": "get_user_api_v1_users__user_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "detail": "Unauthorized",
                        "description": "Unauthorized"
                    },
                    "404": {
                        "detail": "User not found",
                        "description": "Not Found"
                    },
                    "403": {
                        "detail": "Access denied",
                        "description": "Forbidden"
                    },
                    "500": {
                        "detail": "Internal error",
                        "description": "Internal Server Error"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get Users",
                "operationId": "get_users_api_v1_users_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "login",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "",
                            "title": "Login"
                        }
                    },
                    {
                        "name": "first_name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "",
                            "title": "First Name"
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "",
                            "title": "Last Name"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Offset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserResponse"
                                    },
                                    "title": "Response Get Users Api V1 Users Get"
                                }
                            }
                        }
                    },
                    "401": {
                        "detail": "Unauthorized",
                        "description": "Unauthorized"
                    },
                    "403": {
                        "detail": "Access denied",
                        "description": "Forbidden"
                    },
                    "500": {
                        "detail": "Internal error",
                        "description": "Internal Server Error"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "operationId": "login_api_v1_login_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokensResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "detail": "Login unable"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "detail": "Internal error"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/refresh": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Refresh",
                "operationId": "refresh_api_v1_refresh_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshToken"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokensResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "detail": "Refresh token has expired"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "detail": "Internal error"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/{user_id}/folders": {
            "get": {
                "tags": [
                    "core"
                ],
                "summary": "Get Root Folder",
                "operationId": "get_root_folder_api_v1_users__user_id__folders_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Id"
                        }
                    },
                    {
                        "name": "folder_limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Folder Limit"
                        }
                    },
                    {
                        "name": "folder_offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Folder Offset"
                        }
                    },
                    {
                        "name": "file_limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "File Limit"
                        }
                    },
                    {
                        "name": "file_offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "File Offset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FolderResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "detail": "Unauthorized",
                        "description": "Unauthorized"
                    },
                    "403": {
                        "detail": "Access denied",
                        "description": "Forbidden"
                    },
                    "500": {
                        "detail": "Internal error",
                        "description": "Internal Server Error"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "LoginRequest": {
                "properties": {
                    "login": {
                        "type": "string",
                        "title": "Login"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    }
                },
                "type": "object",
                "required": [
                    "login",
                    "password"
                ],
                "title": "LoginRequest"
            },
            "RefreshToken": {
                "properties": {
                    "refresh": {
                        "type": "string",
                        "title": "Refresh"
                    }
                },
                "type": "object",
                "required": [
                    "refresh"
                ],
                "title": "RefreshToken"
            },
            "TokensResponse": {
                "properties": {
                    "access": {
                        "type": "string",
                        "title": "Access"
                    },
                    "refresh": {
                        "type": "string",
                        "title": "Refresh"
                    }
                },
                "type": "object",
                "required": [
                    "access",
                    "refresh"
                ],
                "title": "TokensResponse"
            },
            "UserCreate": {
                "properties": {
                    "login": {
                        "type": "string",
                        "title": "Login"
                    },
                    "first_name": {
                        "type": "string",
                        "title": "First Name"
                    },
                    "last_name": {
                        "type": "string",
                        "title": "Last Name"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    }
                },
                "type": "object",
                "required": [
                    "login",
                    "first_name",
                    "last_name",
                    "email",
                    "password"
                ],
                "title": "UserCreate"
            },
            "UserResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "first_name": {
                        "type": "string",
                        "title": "First Name"
                    },
                    "last_name": {
                        "type": "string",
                        "title": "Last Name"
                    },
                    "login": {
                        "type": "string",
                        "title": "Login"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "is_admin": {
                        "type": "boolean",
                        "title": "Is Admin"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "first_name",
                    "last_name",
                    "login",
                    "email",
                    "is_admin",
                    "created_at"
                ],
                "title": "UserResponse"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}